规范菜谱评论代码

+++ ppeng-module-recipe/src/main/java/fun/zhub/ppeng/mapper/RecipeCommentMapper.java
@@ -4,7 +4,6 @@ import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import fun.zhub.ppeng.entity.RecipeComment;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-import org.apache.ibatis.annotations.Select;
 
 import java.util.List;
 
@@ -21,17 +20,17 @@ public interface RecipeCommentMapper extends BaseMapper<RecipeComment> {
 
     /**
      * 查根据父id询所有的评论信息
-     * @param parentId
-     * @return
+     * @param parentId parentId
+     * @return  List
      */
-    List<RecipeComment> findByParentId(Long parentId);
+    List<RecipeComment> findByParentId(Integer parentId);
 
     /**
      * 通过菜谱id查询该id下的所有parent_id=0的评论信息
-     * @param parentId
-     * @return
+     * @param parentId parentId
+     * @return  List
      */
-    List<RecipeComment> findByRecipeId(@Param("recipeId") Long recipeId,@Param("parentId") Long parentId);
+    List<RecipeComment> findByRecipeId(@Param("recipeId") Long recipeId,@Param("parentId") Integer parentId);
 
 
 }
